scripts upd

+++ scripts/build.sh
@@ -1,9 +1,19 @@
-#!/bin/sh
+#!/usr/bin/env sh
 
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-cd $SCRIPT_DIR/..
+set -e
 
-source ./scripts/_variables.sh
+cd $(dirname $0)
 
-rm ./${BINARY}
+if [ ! -f "$(basename $0)" ]; then
+  echo "Incorrect location"
+  exit 1
+fi
+
+source ./_variables.sh
+cd ..
+
+if [ -f "./${BINARY}" ]; then
+    rm ./${BINARY}
+fi
+
 go build -trimpath -o ${BINARY}
